Add files via upload
Legal witness statements are useful for Natural Language Processing experimentation as they often describe a restricted domain of people and events. 

For this exploration of NLP and in particular the Spacy.io library I chose a 50 page witness statement from an employment tribunal. This was cleaned, taking care to annotate the structure for future reference, then parsed to identify Named Entities using the large Spacy English model. 

The resulting Python dictionary was then used to generate a force directed graph (Gephi / Neo4j) to aid understanding of the statement narrative.